client/tools/client/requirements.txt
# =============================================================================
# 跨平台Python客户端依赖清单
# =============================================================================
# 
# 使用说明:
#   Windows: pip install -r requirements.txt
#   Linux/macOS: pip3 install -r requirements.txt
#
# 注意事项:
#   - Windows用户需要额外安装 pywin32 和 wmi
#   - 使用 pip install -r requirements_windows.txt 安装Windows专属依赖
# =============================================================================

# 核心依赖（所有平台必需）
psutil>=5.9.0                    # 跨平台硬件信息采集，支持Apple Silicon
pycryptodome>=3.18.0             # AES加密，轻量级实现
requests>=2.28.0                 # HTTP客户端（可选，无则使用urllib兜底）

# =============================================================================
# Windows专属依赖（仅Windows需要）
# 请在Windows上额外执行: pip install pywin32 wmi
# =============================================================================
# pywin32>=306                   # Windows API访问
# wmi>=1.5.1                     # WMI查询
